+++ my-vue-app/src/models/ProductFeatures.js
